Output compatibility improvements
Augmented the existing toolbar display output to use the native EE _display method but added some overrides that will ensure the output profiler and the template debugger are not inserted into the final output.

This should allow full compatibility with complex addons that deal with caching and template output (such as Stash).