Fixed extension settings bug
Due to the CI Hooks technique used to grab the final display output a bug was
encountered whereby the extension settings were not available (in this case the
selected theme was not present).

This is due to the Ee_debug_toolbar_ext being instantiated twice:

* By ExpressionEngine on the session_end EE hook.
* By CodeIgniter on the display_override CI hook

The EE settings are only provided on the first initialisation by EE. This problem was
solved by creating a static property to persist the settings between instantiations.