Fixes eedt.js AJAX communications
Now any method can communicate live with the server from js by
using the eedt.ajax() method. The call is routed through a central
EEDT method that initialised the specified class and calls the
specified method.  This is done in order to prevent the need for
creating an EE model just for ACT communication.

The method allows for callbacks as an argument, but it also returns a
deferred object:

    //Callback
    eedt.ajax('Class_name', "method_name", function(data){
        //Code
    });

    //Deferred
    eedt.ajax('Class_name', "method_name").then(function(data){
        //Code
    });