Adds eedt.ready() method
This method is designed to be used with JS scripts that are registered
to be loaded on page load.  By loading on page load these scripts run
the risk of loading before jQuery or eedt.js is ready (leading to
JS errors).

The eedt.ready() accepts a callback as an argument and allows
developers to be able to have this callback triggered at a
point when jQuery is guaranteed to be on the page and also when the
eedt.js API is fully loaded.

The callback function recieves two arguments when triggered: the
jQuery object and the eedt.js library itself.

Usage:

eedt.js(function($, eedt){
    //code
})